requirements.txt
cv2-python
pyttsx3
pytesseract
pillow
